increased fingerprint from 32-bits to 48-bits and refactored
rational: decrease risk of collisions though this only becomes a potential
problem with a large number of keys.

For every order of magnitude increase in the number of secrets, the number of
collisions increases by two orders of magnitude.

For example in practice a 32-bit fingerprint has about:

- 1.2 collisions every 100,000 secrets.
- 120 collisions every 1 million secrets.
- 12,000 collisions every 10 million secrets.
- 1,200,000 collisions every 100 million secrets.

By increasing from 32-bit to 48-bit we stave off collisions in practice beyond
the realm that we will probably every encounter.

With a 48-bit fingerprint I ran a simple simulation and couldn't get even 1
conflict in 10,000,000 keys.