force GnuPG to use the AES-128 cipher by default
GnuPG defaults to CAST5 because this is one of the strongest royalty-free
ciphers available before AES, which provides interoperability with older
versions of PGP.

But CAST5 sucks as a default because:

1) Bruce Schneier said so. Regarding it's design process he said "yuk".
2) It's pretty obscure as ciphers go. AES is more comforting.

So we're going with AES, but which AES? There is AES-128 (the default),
AES-192, and AES-256.

There are no know attacks against AES-128, but since AES has known
vulnerabilities related to large keys there are attacks that reduce the AES-256
to 2^99, which actually makes it even weaker than AES-128. 192 is similarly
effected but less so.

So now I'm suspicious of any key size larger than the default AES was designed
for. AES-128 it is!