bugfix: save all mode bits in fsdelta + don't assume non-existent status path is a directory
Format change:

This changes the format of status changes in /TKLBAM/fsdelta to store all
stat.st_mode bits.

Backwards compatibility: we should still be able to restore old backups created
using the previous incorrect behavior , but the bug is only fixed if you
restore from a backup created or updated after this fix. In other words, old
backups are restored using the old logic.

Forward compatibility: old versions of TKLBAM will restore new backups using
the old logic. The log output for chmod status changes will look strange
because all mode bits are passed to chmod. Functionally, chmod will just ignore
the extra extra bits.

Bug explanation:

Previously if you changed permissions or ownership of a path in a backup and
restored the backup to a system that doesn't have this path, tklbam would
assume the path was a directory and create it.

This is absolutely the wrong thing to do. There's no guarantee that paths exist
on the restored system. For example, if you change permission on a path and
then restore that backup to a newer version of the appliance where that path
doesn't exist creating a new directory will makes things worse. The correct
thing to do is to ignore the status change.