extended hooks mechanism to support signed profile-level hooks
rational: profile level hooks could be used to implement dynamic appliance
specific restore and backup logic.

For example, this could be useful in applying automatic fixes to make it easier
to restore the backup from an old version of an appliance on a new version of
the appliance (or vice versa).